Bugfixes - Values back to float

+++ Assets/Scripts/OpenBCI_Sample.cs
@@ -4,16 +4,16 @@ using System.Collections;
 public class OpenBCI_Sample{
 
 	public short sampleId; // The id of the sample
-	public int[] channelSample; // The data of all channels, either 8 or 16
-	public short[] accelData; // The data of the accelerometer, 3 directions
+	public float[] channelSample; // The data of all channels, either 8 or 16
+	public float[] accelData; // The data of the accelerometer, 3 directions
 
 	public OpenBCI_Sample(){ // Constructor with default values
 		sampleId = -1;
-		channelSample = new int[8];
-		accelData = new short[3];
+		channelSample = new float[8];
+		accelData = new float[3];
 	}
 
-	public OpenBCI_Sample(short sampleId, int[] channelSample, short[]accelData){ // Constructor with given values
+	public OpenBCI_Sample(short sampleId, float[] channelSample, float[]accelData){ // Constructor with given values
 		this.sampleId = sampleId;
 		this.channelSample = channelSample;
 		this.accelData = accelData;
